+++ mocks/mock_test.go
@@ -17,6 +17,7 @@ import (
 func Test_MockClient(t *testing.T) {
 	testWorkflowID := "test-workflowid"
 	testRunID := "test-runid"
+	testFirstExecutionRunID := "test-first-execution-runid"
 	testWorkflowName := "workflow"
 	testWorkflowInput := "input"
 	mockClient := &Client{}
@@ -24,6 +25,7 @@ func Test_MockClient(t *testing.T) {
 	mockWorkflowRun := &WorkflowRun{}
 	mockWorkflowRun.On("GetID").Return(testWorkflowID).Times(4)
 	mockWorkflowRun.On("GetRunID").Return(testRunID).Times(4)
+	mockWorkflowRun.On("GetFirstExecutionRunID").Return(testFirstExecutionRunID).Times(4)
 	mockWorkflowRun.On("Get", mock.Anything, mock.Anything).Return(nil).Times(2)
 
 	mockClient.On("ExecuteWorkflow", mock.Anything, mock.Anything, mock.Anything, mock.Anything).Return(mockWorkflowRun, nil).Once()
@@ -32,6 +34,7 @@ func Test_MockClient(t *testing.T) {
 	require.NoError(t, err)
 	require.Equal(t, testWorkflowID, wr.GetID())
 	require.Equal(t, testRunID, wr.GetRunID())
+	require.Equal(t, testFirstExecutionRunID, wr.GetFirstExecutionRunID())
 	require.NoError(t, mockWorkflowRun.Get(context.Background(), &testWorkflowID))
 
 	mockClient.On("SignalWithStartWorkflow", mock.Anything, mock.Anything, mock.Anything, mock.Anything, mock.Anything, mock.Anything, mock.Anything).Return(mockWorkflowRun, nil).Once()
@@ -40,6 +43,7 @@ func Test_MockClient(t *testing.T) {
 	require.NoError(t, err)
 	require.Equal(t, testWorkflowID, wr.GetID())
 	require.Equal(t, testRunID, wr.GetRunID())
+	require.Equal(t, testFirstExecutionRunID, wr.GetFirstExecutionRunID())
 
 	mockClient.On("CancelWorkflow", mock.Anything, testWorkflowID, testRunID).Return(nil).Once()
 	err = mockClient.CancelWorkflow(context.Background(), testWorkflowID, testRunID)
@@ -47,12 +51,14 @@ func Test_MockClient(t *testing.T) {
 	require.NoError(t, err)
 	require.Equal(t, testWorkflowID, wr.GetID())
 	require.Equal(t, testRunID, wr.GetRunID())
+	require.Equal(t, testFirstExecutionRunID, wr.GetFirstExecutionRunID())
 
 	mockClient.On("GetWorkflow", mock.Anything, testWorkflowID, testRunID).Return(mockWorkflowRun).Once()
 	wr = mockClient.GetWorkflow(context.Background(), testWorkflowID, testRunID)
 	mockClient.AssertExpectations(t)
 	require.Equal(t, testWorkflowID, wr.GetID())
 	require.Equal(t, testRunID, wr.GetRunID())
+	require.Equal(t, testFirstExecutionRunID, wr.GetFirstExecutionRunID())
 	require.NoError(t, wr.Get(context.Background(), &testWorkflowID))
 
 	mockHistoryIter := &HistoryEventIterator{}